projects/deep_learning_devex/pyproject.toml
[tool.poetry]
name = "dl-dep-mgr"
version = "0.1.0"
description = "CLIP fine-tuning pipeline with unified local/Databricks/serving experience"
authors = ["Eli Swanson <eli.swanson@databricks.com>"]
readme = "README.md"
packages = [{ include = "dl_dep_mgr", from = "src" }]
license = "DB License"

[tool.poetry.dependencies]
# Not necessary to pin all versions in pyproject.toml 
# transitive dependencies will be pinned by Poetry, and exported for DBR + serving.
python = "^3.11,<3.13" # DBR 17.3 uses Python 3.12
# Core ML - pinned to DBR 17.3 ML GPU versions
torch = "2.7.0"
torchvision = "0.22.0"
# CLIP and streaming
open-clip-torch = "^2.24.0"
mosaicml-streaming = "^0.10.0"
# MLflow and Databricks - DBR 17.3 versions
mlflow = "^3.0.0"
# Databricks SDK is pinned to avoid version-marker conflicts.
# For now, recommend keeping versions between 0.49 and 0.68.
# Avoiding some volume and mlflow issues.
databricks-sdk = "0.49.0"
# Utilities
pillow = "11.1.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
# Pin to avoid version-marker conflicts in requirements.txt
contourpy = "1.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
black = "^24.8.0"
ruff = "^0.6.0"

[tool.poetry.scripts]
train = "dl_dep_mgr.train:main"
deploy_serving = "dl_dep_mgr.deploy_serving:main"
test_inference = "dl_dep_mgr.test_inference:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--cov=src/dl_dep_mgr", "--cov-report=term-missing"]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
